feat(security): resolved dependency cycle [PW-305]
