--- backend/java/src/main/java/com/michaelyi/personalwebsite/auth/AuthPasswordEncoderConfig.java
@@ -1,14 +1,0 @@
-package com.michaelyi.personalwebsite.auth;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@Configuration
-public class AuthPasswordEncoderConfig {
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-}
